impl highest_ram_addr_in_use()
