+++ src/elf.rs
@@ -247,15 +247,20 @@ fn extract_symbols(elf: &ElfFile) -> anyhow::Result<(Option<u32>, /* uses heap: 
 pub(crate) struct TargetInfo {
     pub(crate) target: Target,
     pub(crate) active_ram_region: Option<RamRegion>,
+    pub(crate) highest_ram_addr_in_use: Option<u32>, // todo maybe merge
 }
 
 impl TargetInfo {
-    pub(crate) fn new(chip: &str, initial_sp: u32) -> anyhow::Result<Self> {
+    pub(crate) fn new(chip: &str, elf: &ProcessedElf) -> anyhow::Result<Self> {
         let target = probe_rs::config::registry::get_target_by_name(chip)?;
-        let active_ram_region = extract_active_ram_region(&target, initial_sp);
+        let active_ram_region =
+            extract_active_ram_region(&target, elf.vector_table.initial_stack_pointer);
+        let highest_ram_addr_in_use =
+            extract_highest_ram_addr_in_use(elf, active_ram_region.as_ref())?;
         Ok(Self {
             target,
             active_ram_region,
+            highest_ram_addr_in_use,
         })
     }
 }
@@ -279,6 +284,33 @@ fn extract_active_ram_region(target: &Target, initial_sp: u32) -> Option<RamRegi
         })
         .next()
         .cloned()
+}
+
+fn extract_highest_ram_addr_in_use(
+    elf: &ElfFile,
+    active_ram_region: Option<&RamRegion>,
+) -> anyhow::Result<Option<u32>> {
+    let mut highest_ram_addr_in_use = None;
+    for sect in elf.sections() {
+        // If this section resides in RAM, track the highest RAM address in use.
+        if let Some(ram) = active_ram_region {
+            if sect.size() != 0 {
+                let last_addr = sect.address() + sect.size() - 1;
+                let last_addr = last_addr.try_into()?;
+                if ram.range.contains(&last_addr) {
+                    log::debug!(
+                        "section `{}` is in RAM at 0x{:08X}-0x{:08X}",
+                        sect.name().unwrap_or("<unknown>"),
+                        sect.address(),
+                        last_addr,
+                    );
+                    highest_ram_addr_in_use = highest_ram_addr_in_use.max(Some(last_addr));
+                }
+            }
+        }
+    }
+
+    Ok(highest_ram_addr_in_use)
 }
 
 // obtained via probe-rs?
